Implement challenge 3 output

+++ src/estimator.py
@@ -33,8 +33,8 @@ def estimator(data):
 
     avgDailyIncomeInUSD = data['region']['avgDailyIncomeInUSD']
     avgDailyIncomePopulation = data['region']['avgDailyIncomePopulation']
-    dollarsInFlightC = (infectionsByRequestedTimeC * avgDailyIncomePopulation * 100 * avgDailyIncomeInUSD * float(days)) / 100
-    dollarsInFlightS = (infectionsByRequestedTimeS * avgDailyIncomePopulation * 100 * avgDailyIncomeInUSD * float(days)) / 100
+    dollarsInFlightC = int((infectionsByRequestedTimeC * avgDailyIncomePopulation * 100 * avgDailyIncomeInUSD * float(days)) / 100)
+    dollarsInFlightS = int((infectionsByRequestedTimeS * avgDailyIncomePopulation * 100 * avgDailyIncomeInUSD * float(days)) / 100)
